embed hljs + work around commonjs pitfalls
i chose to copy the ruby language in. technically, if whoever uses this has a good babel config this step wouldnt be necessary. however, this is a fool-proof way that doesnt depend as much on the user config and simply transforms the commonjs syntax to the ruby.min they offer in their cdn:

https://cdnjs.cloudflare.com/ajax/libs/highlight.js/9.12.0/languages/ruby.min.js